# schemas/api/openapi.yaml
openapi: 3.0.3
info:
  title: DIDHub API
  version: 0.1.0
  description: >-
    HTTP API surface for the DIDHub backend. This schema intentionally focuses on
    contracts needed by the generator stubs and TypeScript client.
servers:
  - url: https://api.didhub.local
    description: Placeholder server URL
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          default: 1
        perPage:
          type: integer
          minimum: 1
          default: 20
        total:
          type: integer
          minimum: 0
      required:
        - page
        - perPage
        - total
...
    Alter:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        pronouns:
          type: string
        description:
          type: string
        age:
          type: string
          description: Age of the alter
        gender:
          type: string
          description: Gender identity of the alter
        birthday:
          type: string
          description: Birthday of the alter
        sexuality:
          type: string
          description: Sexual orientation of the alter
        species:
          type: string
          description: Species of the alter
        alterType:
          type: string
          description: Type of alter (e.g., fictive, introject, etc.)
        job:
          type: string
          description: Occupation or role of the alter
        weapon:
          type: string
          description: Associated weapon or tool
        systemRoles:
          type: array
          items:
            type: string
          description: Roles within the system
        isSystemHost:
          type: boolean
          description: Whether this alter is a system host
        isDormant:
          type: boolean
          description: Whether the alter is dormant or dead
        isMerged:
          type: boolean
          description: Whether the alter has merged with another
        soulSongs:
          type: array
          items:
            type: string
          description: Songs associated with this alter
        interests:
          type: array
          items:
            type: string
          description: Interests and hobbies
        triggers:
          type: array
          items:
            type: string
          description: Trigger warnings
        notes:
          type: string
          description: Long-form notes about the alter
        primaryUploadId:
          type: string
          format: uuid
          description: >-
            Optional reference to the primary upload associated with this alter
            (use GET /files/{fileId} to retrieve file data).
        systemId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - systemId
        - createdAt
        - updatedAt
    PaginatedAlterResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Alter'
        pagination:
          $ref: '#/components/schemas/Pagination'
      required:
        - items
        - pagination
    CreateAlterRequest:
      type: object
      properties:
        name:
          type: string
        pronouns:
          type: string
        description:
          type: string
        age:
          type: string
        gender:
          type: string
        birthday:
          type: string
        sexuality:
          type: string
        species:
          type: string
        alterType:
          type: string
        job:
          type: string
        weapon:
          type: string
        systemRoles:
          type: array
          items:
            type: string
        isSystemHost:
          type: boolean
        isDormant:
          type: boolean
        isMerged:
          type: boolean
        soulSongs:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        triggers:
          type: array
          items:
            type: string
        notes:
          type: string
      required:
        - name
    UpdateAlterRequest:
      type: object
      properties:
        name:
          type: string
        pronouns:
          type: string
        description:
          type: string
        age:
          type: string
        gender:
          type: string
        birthday:
          type: string
        sexuality:
          type: string
        species:
          type: string
        alterType:
          type: string
        job:
          type: string
        weapon:
          type: string
        systemRoles:
          type: array
          items:
            type: string
        isSystemHost:
          type: boolean
        isDormant:
          type: boolean
        isMerged:
          type: boolean
        soulSongs:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        triggers:
          type: array
          items:
            type: string
        notes:
          type: string
        images:
          type: array
          items:
            type: string
            format: uuid
...
    Subsystem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
...
    SubsystemMemberRequest:
      type: object
      properties:
        alterId:
          type: string
          format: uuid
      required:
        - alterId
...
paths:
...
  /alters:
    get:
      tags: [Alters]
      summary: List alters
      operationId: listAlters
      x-handler:
        delegate: crate::handlers::alters::list::list
        passHeaders: true
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: perPage
          in: query
          schema:
            type: integer
            minimum: 1
        - name: search
          in: query
          schema:
            type: string
        - name: subsystemId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Paginated alters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAlterResponse'
      security:
        - bearerAuth: []
    post:
      tags: [Alters]
      summary: Create alter
      operationId: createAlter
      x-handler:
        delegate: crate::handlers::alters::create::create
        passHeaders: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlterRequest'
      responses:
        '201':
          description: Created alter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alter'
      security:
        - bearerAuth: []
...
  /alters/{alterId}:
    parameters:
      - name: alterId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Alters]
      summary: Get alter
      operationId: getAlter
      x-handler:
        delegate: crate::handlers::alters::get::get
        passHeaders: true
      responses:
        '200':
          description: Alter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alter'
      security:
        - bearerAuth: []
    patch:
      tags: [Alters]
      summary: Update alter
      operationId: updateAlter
      x-handler:
        delegate: crate::handlers::alters::update::update
        passHeaders: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlterRequest'
      responses:
        '200':
          description: Updated alter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alter'
      security:
        - bearerAuth: []
    delete:
      tags: [Alters]
      summary: Delete alter
      operationId: deleteAlter
      x-handler:
        delegate: crate::handlers::alters::delete::delete
        passHeaders: true
      responses:
        '204':
          description: Alter deleted
      security:
        - bearerAuth: []
...
  /alters/{alterId}/subsystem:
    get:
      tags: [Alters]
      summary: Get alter subsystem
      operationId: getAlterSubsystem
      x-handler:
        delegate: crate::handlers::alters::subsystem::get
        passHeaders: true
      parameters:
        - name: alterId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Subsystem
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subsystem'
      security:
        - bearerAuth: []
    put:
      tags: [Alters]
      summary: Set alter subsystem
      operationId: setAlterSubsystem
      x-handler:
        delegate: crate::handlers::alters::subsystem::set
        passHeaders: true
      parameters:
        - name: alterId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubsystemMemberRequest'
      responses:
        '200':
          description: Updated subsystem
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subsystem'
      security:
        - bearerAuth: []
    delete:
      tags: [Alters]
      summary: Remove alter subsystem
      operationId: removeAlterSubsystem
      x-handler:
        delegate: crate::handlers::alters::subsystem::delete
        passHeaders: true
      parameters:
        - name: alterId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Subsystem cleared
      security:
        - bearerAuth: []
...
security:
  - bearerAuth: []